fix iso date parsing
With this patch applied, you should be able to do:

    kubectl logs --timestamps  | ts -r

and receive relative timestamps in your output.

---

Review by Joey: This fixes the year to have 4 digits,
which seems obviously right.

It also allows a trailing Z, for dates in UTC.
Other dates already matched by this regexp can
have time zones, and ts just ignores time zones
and assumes it's in the local time zone. I don't know
if that's really a good idea, but accepting "Z"
does not open a new can of worms.